Limit pin internal pull-resistors enabled. Re-wrote read_double() function. Correctly changed all 'double's to 'float's.
- Limit pin internal pull-resistors now enabled. Normal high operation.
This will be the standard going forward.

- Updated all of the 'double' variable types to 'float' to reflect what
happens when compiled for the Arduino. Also done for compatibility
reasons to @jgeisler0303 's Grbl simulator code.

- G-code parser will now ignore 'E' exponent values, since they are
reserved g-code characters for some machines. Thanks @csdexter!

- The read_double() function was re-written and optimized for use in
Grbl. The strtod() avr lib was removed.

+++ main.c
@@ -62,7 +62,7 @@ int main(void)
       // releases will auto-reset the machine position back to [0,0,0] if an abort is used while 
       // grbl is moving the machine.
       int32_t last_position[3];
-      double last_coord_system[N_COORDINATE_SYSTEM][3];
+      float last_coord_system[N_COORDINATE_SYSTEM][3];
       memcpy(last_position, sys.position, sizeof(sys.position)); // last_position[] = sys.position[]
       memcpy(last_coord_system, sys.coord_system, sizeof(sys.coord_system)); // last_coord_system[] = sys.coord_system[]
